skip `ssl` import if not available
`utils.py` has support for detecting whether the `ssl` module is available, and
we can use this to omit SSL-specific funcionality while still providing other
features (e.g. unencrypted connections).

Prior to this patch, the `connection.py` modules both triggered an `ImportError`
due to unconditional imports of the `ssl` module.  Now, we check
`utils.SSL_AVAILABLE` prior to attempting the import and only raise an error
later if (and only if) the application requests an encrypted connection.  This
helps support platforms such as `wasm32-wasi` where the `ssl` module is not
built by default.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>